+++ generator/codegen/templates/entities.go.tmpl
@@ -12,7 +12,7 @@ import (
 
 {{ range .Collections }}
 // ============================================================================================
-// This extends {{ .ItemRef }} and {{ .CollectionRef }} with pagination methods
+// This extends {{ .ItemRef }} and {{ .CollectionRef }} definitions
 // ============================================================================================
 
 // FetchType returns the resource type
@@ -142,7 +142,7 @@ func New{{ .CollectionRef }}Collection() IStaticPage {
 {{ end }}
 {{ range .MessageCollections -}}
 // ============================================================================================
-// This extends {{ .ItemRef }} with pagination methods
+// This extends {{ .ItemRef }} definitions
 // ============================================================================================
 
 // FetchType returns the resource type
@@ -184,7 +184,7 @@ func New{{ .IteratorRef }}(elements []{{ .ItemRef }}) (IIterator, error) {
 {{ end }}
 {{ range .NotificationFeedCollections -}}
 // ============================================================================================
-// This extends {{ .ItemRef }} with pagination methods
+// This extends {{ .ItemRef }} definitions
 // ============================================================================================
 
 // FetchType returns the resource type
