+++ generator/codegen/types.go
@@ -1,5 +1,11 @@
 package codegen
 
+import (
+	validation "github.com/go-ozzo/ozzo-validation"
+
+	configUtils "github.com/ARM-software/golang-utils/utils/config"
+)
+
 type Collection struct {
 	CollectionRef string
 	ItemRef       string
@@ -26,4 +32,38 @@ type CollectionParams = struct {
 	Collections
 	MessageCollections
 	NotificationFeedCollections
+}
+
+type Data struct {
+	Params          any
+	SpecPath        string
+	TemplatePath    string
+	DestinationPath string
+}
+
+type ExtensionsConfig struct {
+	Input    string `mapstructure:"input"`
+	Output   string `mapstructure:"output"`
+	Template string `mapstructure:"template"`
+}
+
+func DefaultExtensionsConfig() *ExtensionsConfig {
+	return &ExtensionsConfig{
+		Input:    "",
+		Output:   "",
+		Template: "templates/entities.go.tmpl",
+	}
+}
+
+func (cfg *ExtensionsConfig) Validate() error {
+	err := configUtils.ValidateEmbedded(cfg)
+	if err != nil {
+		return err
+	}
+
+	return validation.ValidateStruct(cfg,
+		validation.Field(&cfg.Input, validation.Required),
+		validation.Field(&cfg.Output, validation.Required),
+		validation.Field(&cfg.Template, validation.Required),
+	)
 }
